chore: update test for ViewModel
