+++ app/build.gradle
@@ -44,9 +44,9 @@ android {
         }
         debug {
             testCoverageEnabled = true
-            buildConfigField "String", "DEFAULT_BASE_URL", '"https://open-event-api-dev.herokuapp.com/v1/"'
+            buildConfigField "String", "DEFAULT_BASE_URL", '"https://api.eventyay.com/v1/"'
             buildConfigField 'Boolean', 'HIDE_DRAWER_ITEMS', 'false'
-            resValue "string", "default_base_url", '"https://open-event-api-dev.herokuapp.com/v1/"'
+            resValue "string", "default_base_url", '"https://api.eventyay.com/v1/"'
         }
     }
     dataBinding {
